Commit is weird
Just trying to make sure i know how to commit changes and stuff correctly
